% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich_diff.R
\name{rich_diff}
\alias{rich_diff}
\title{Get Rich Diffs for Rmds}
\usage{
rich_diff(gh_file_path)
}
\arguments{
\item{gh_file_path}{path to the file to compare. Eg a .rmd or .qmd}
}
\value{
Viewer opens with diff highlighting individual word differences.
}
\description{
Modified from
https://gist.github.com/benmarwick/85e853b258a653f0b44d2df293fba222. See rich
diffs of two commits of a single R Markdown document on a GitHub repo. Will
work on whatever branch is currently checked out.
}
